refactored sharding
