+++ src/main/java/io/anserini/collection/C4Collection.java
@@ -27,6 +27,7 @@ import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
+import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 import java.util.zip.GZIPInputStream;
@@ -57,6 +58,21 @@ public class C4Collection extends DocumentCollection<C4Collection.Document> {
         return 0;
       return character;
     }
+  }
+
+  private int getFileNumber(String fileName) {
+    try {
+      int fileNumStart = fileName.indexOf("c4-train.") + 9;
+      return Integer.parseInt(fileName.substring(fileNumStart, fileNumStart + 5));
+    } catch (final NumberFormatException e) {
+      return fileName.hashCode();
+    }
+  }
+
+  @Override
+  public List<Path> getSegmentPaths(int shardCount, int currShard) {
+    List<Path> segments = super.getSegmentPaths();
+    return segments.stream().filter(x -> getFileNumber(x.toString()) % shardCount == currShard).collect(Collectors.toList());
   }
 
   public static class Segment extends FileSegment<C4Collection.Document>{
